requirements.txt
pandas
numpy
librosa
soundfile
openai-whisper
torch
torchaudio
spacy
scikit-learn
